[rundroid] Honor `adb shell` exit code
`adb shell` doesn't return the exit code of the program that ran on the device
(see: https://code.google.com/p/android/issues/detail?id=3254).

As a result, tests that abort are marked as successful. In the best case
scenario, this leads to an "unexpected pass", which informs SwiftAndroid
developers that something is up. In the worst case, a test that is expected
to pass, but actually fails on device, is marked as a passing test--this
undermines trust in the test suite.

Use the exit code from `adb shell` and fail the test if it is anything but
`0`. This change reveals four additional failing tests for SwiftAndroid.